ref: reorganizing SDL2 dependencies and enemies
