+++ Source/Actors/Actor.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <vector>
-#include <SDL2/SDL_stdinc.h>
+#include <SDL_stdinc.h>
 #include "../Components/ColliderComponents/AABBColliderComponent.h"
 
 enum class ActorState {
