+++ Source/Components/Component.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <SDL2/SDL_stdinc.h>
+#include <SDL_stdinc.h>
 
 class Component {
     public:
