+++ Source/Systems/Audio/AudioSystem.cpp
@@ -1,6 +1,6 @@
 #include "AudioSystem.h"
-#include "SDL2/SDL.h"
-#include "SDL2/SDL_mixer.h"
+#include "SDL.h"
+#include "SDL_mixer.h"
 #include <filesystem>
 
 SoundHandle SoundHandle::Invalid;
