+++ Source/Systems/Audio/AudioSystem.h
@@ -3,7 +3,7 @@
 #include <map>
 #include <string>
 #include <vector>
-#include "SDL2/SDL_stdinc.h"
+#include "SDL_stdinc.h"
 
 class SoundHandle {
     public:
