+++ Source/UIElements/UIElement.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../Utils/Math.h"
-#include <SDL2/SDL.h>
+#include <SDL.h>
 
 class UIElement {
 public:
