+++ Source/UIElements/UIText.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <string>
-#include <SDL2/SDL.h>
+#include <SDL.h>
 #include "../Utils/Math.h"
 #include "UIElement.h"
